+++ backend/token.go
@@ -1,8 +1,12 @@
 package main
 
 import (
+	"fmt"
 	"github.com/bcfoodapp/streetfoodlove/uuid"
+	"github.com/gin-gonic/gin"
 	"github.com/golang-jwt/jwt"
+	"net/http"
+	"strings"
 	"time"
 )
 
@@ -10,6 +14,48 @@ import (
 type AccessTokenClaims struct {
 	UserID uuid.UUID
 	jwt.StandardClaims
+}
+
+var tokenSecret = []byte("key")
+
+var userIDKey = "userID"
+
+// GetToken is a middleware to validate token and get user ID. The user ID is retrieved with
+// getTokenFromContext().
+func GetToken(c *gin.Context) {
+	headerValue := c.GetHeader("Authorization")
+	const bearer = "Bearer "
+	if !strings.HasPrefix(headerValue, bearer) {
+		c.AbortWithError(http.StatusBadRequest, fmt.Errorf("token not in bearer field"))
+		return
+	}
+
+	tokenStr := strings.TrimPrefix(headerValue, bearer)
+
+	claims := &AccessTokenClaims{}
+	_, err := jwt.ParseWithClaims(tokenStr, claims, func(token *jwt.Token) (interface{}, error) {
+		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
+			return nil, fmt.Errorf("invalid signing method")
+		}
+
+		return tokenSecret, nil
+	})
+
+	if err != nil {
+		c.AbortWithError(http.StatusBadRequest, err)
+		return
+	}
+
+	if claims.Valid() != nil {
+		c.AbortWithError(http.StatusUnauthorized, claims.Valid())
+		return
+	}
+
+	c.Set(userIDKey, claims.UserID)
+}
+
+func getTokenFromContext(c *gin.Context) uuid.UUID {
+	return c.MustGet(userIDKey).(uuid.UUID)
 }
 
 func generateAccessToken(userID uuid.UUID) string {
@@ -32,6 +78,23 @@ func generateAccessToken(userID uuid.UUID) string {
 type RefreshTokenClaims struct {
 	GoogleID string
 	jwt.StandardClaims
+}
+
+func validateRefreshToken(refreshToken string) (*RefreshTokenClaims, error) {
+	claims := &RefreshTokenClaims{}
+	_, err := jwt.ParseWithClaims(refreshToken, claims, func(token *jwt.Token) (interface{}, error) {
+		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
+			return nil, fmt.Errorf("invalid signing method")
+		}
+
+		return tokenSecret, nil
+	})
+
+	if err != nil {
+		return nil, err
+	}
+
+	return claims, claims.Valid()
 }
 
 func generateRefreshToken(googleID string) string {
